Intern_Submissions/Priyanshu_Dewangan/requirements.txt
# Podcast Insights AI - Complete Requirements
# Includes all dependencies for: app.py, audio_processor.py, rag_system.py, 
# utils.py, api_server.py, and podcast_summarizer.ipynb

# Web Framework
streamlit==1.28.1

# Audio Processing
librosa==0.10.0
pydub==0.25.1
openai-whisper==20231117

# Deep Learning & ML
torch==2.1.2
torchvision==0.16.2
torchaudio==2.1.2
transformers==4.35.2
sentence-transformers==2.3.1
scikit-learn==1.3.2

# Data Processing
numpy==1.24.3
pandas==2.1.3
scipy==1.11.4

# Visualization
matplotlib==3.8.2

# NLP
nltk==3.8.1

# Vector Database
chromadb==0.4.24

# Graph Processing
networkx==3.2.1

# API Server (Optional)
fastapi==0.104.1
uvicorn==0.24.0

# JSON handling (standard library, included for clarity)
# json - built-in

# Additional utilities
joblib>=1.3.0  # For model persistence
regex>=2023.10.3  # For advanced regex patterns in RAG
